Create RandomWalk.java
As a retort to the recent violent activities carried out by the Mandarin, Tony has fearlessly let out his Malibu address in public. Their (Tony's and Pepper's) place is now bombed by Mandarin's henchmen and Tony wasn't prepared to fight back. The men have kidnapped Pepper and have thrown a challenge at him. They want Tony alive and Pepper is just a hostage to get Tony to the places where Mandarin is hiding. They have smartly infiltrated the city and are based in the undergrounds of either of the three most prominent buildings of the city. As you would believe Whitehouse is more prominent than a bank, which is in turn more prominent than a house. The prominent places have connections to other prominent places and Tony's task is to come down to a list of possible(top 3 most prominent) places where Mandarin could be based.