Create CostEffectiveTree.java
the cost of a binary search tree is defined as its height. You are given a sorted sequence of integers as input. Your task is to create the binary search tree of minimum cost containing the given sequence of numbers. Print the preorder traversal of the minimum cost binary search tree as output.